Move events resource types
