+++ tests/utils/fixtures/mock_event.py
@@ -1,6 +1,6 @@
 import datetime
 
-from workos.resources.events import DirectoryActivatedEvent
+from workos.types.events import DirectoryActivatedEvent
 from workos.types.events.directory_payload_with_legacy_fields import (
     DirectoryPayloadWithLegacyFields,
 )
