+++ workos/events.py
@@ -4,7 +4,7 @@ import workos
 from workos.types.events.list_filters import EventsListFilters
 from workos.typing.sync_or_async import SyncOrAsync
 from workos.utils.request_helper import DEFAULT_LIST_RESPONSE_LIMIT, REQUEST_METHOD_GET
-from workos.resources.events import Event, EventType
+from workos.types.events import Event, EventType
 from workos.utils.http_client import AsyncHTTPClient, SyncHTTPClient
 from workos.utils.validation import Module, validate_settings
 from workos.resources.list import (
