+++ workos/types/directory_sync/__init__.py
@@ -3,4 +3,3 @@ from .directory_state import *
 from .directory_type import *
 from .directory_user import *
 from .directory import *
-from .list_filters import *
